feat(bevy_gltf_components): testing a different approach using only gltf extras (#134)
 * not using / tracking gltf files anymore (should avoid issues with scenes vs gltf files)
 * restructured accordingly
 * closes #102
 * closes #111
